requirements.txt
flask
discord
pandas
pypokedex
pytz
requests
urllib3
python-dotenv
yt_dlp
colorama
pynacl
googletrans==3.1.0-alpha
spotdl